renaming and new screens
