+++ src/App.vue
@@ -1,13 +1,13 @@
 <script setup>
 // This starter template is using Vue 3 <script setup> SFCs
 // Check out https://v3.vuejs.org/api/sfc-script-setup.html#sfc-script-setup
-import MatchMaker from './pages/MatchMaker.vue';
+import TheTournament from "./pages/TheTournament.vue";
+import TheAppContainer from "./pages/TheAppContainer.vue";
 </script>
 
 <template>
-  <MatchMaker />
+  <TheTournament />
+  <TheAppContainer />
 </template>
 
-<style>
-
-</style>
+<style></style>
